Refactor interp
